requirements.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
clarabel==0.11.1
click==8.2.1
curl_cffi==0.13.0
cvxpy==1.7.2
ecos==2.0.14
filelock==3.19.1
frozendict==2.4.6
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.1.10
huggingface-hub==0.35.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.1
MarkupSafe==3.0.2
mpmath==1.3.0
multitasking==0.0.12
narwhals==2.5.0
networkx==3.5
newsapi-python==0.2.7
nsepy==0.8
numpy==2.3.3
osqp==1.0.4
packaging==25.0
pandas==2.3.2
peewee==3.18.2
pillow==11.3.0
platformdirs==4.4.0
plotly==5.24.1
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
pyportfolioopt==1.5.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
scs==3.2.8
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
xgboost==3.0.5
yfinance==0.2.65
